fix(guardrails): fail-closed pipeline hardening + decompose module
Correctness:
- from_compiled_json now returns Result; a present-but-malformed
  guardrails block poisons the policy so every request fails closed
  instead of silently degrading to "no guardrails" (which also
  disarmed the sibling route-gap guard). The loader installs an
  unbuildable sentinel stage on parse error.
- SSE moderation reconstructs multi-byte UTF-8 code points split
  across chunk boundaries, so the scanner sees the same text the
  client receives (no U+FFFD divergence).
- Streaming delta extraction scans every choices[] entry, not just
  choices[0].

Refactor:
- Split guardrails.rs (>1600 lines) into guardrails/{mod,backend,
  stream,tests}.rs, each under the 800-line CI cap. Public API is
  unchanged (re-exported); a few items raised to pub(crate) so the
  test module can construct and inspect them.

Regression tests for the UTF-8 split (flagged and benign),
multi-choice extraction, and malformed-config fail-closed.
